refactor: move out update logic
